pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agno-2-trials"
version = "0.1.0"
description = "Basic FastHTML hello world project"
readme = {file = "README.org", content-type = "text/plain"}
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pip",
    "python-fasthtml",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.hatch.build.targets.wheel]
packages = ["."]
